# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: HausHaltsMeister API
  version: 0.1.0
  description: Ledger-based finance API with budgeting, investments, and credit card flows.
servers:
  - url: /
...
security:
  - BearerAuth: []
paths:
...
  /ledgers:
    get:
      tags: [Ledgers]
      summary: List ledgers
      responses:
        '200':
          description: Ledgers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerWithRole'
    post:
      tags: [Ledgers]
      summary: Create ledger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LedgerCreateRequest'
      responses:
        '201':
          description: Ledger
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ledger'
...
  /ledgers/{ledgerId}/accounts:
    get:
      tags: [Accounts]
      summary: List accounts
      parameters:
        - $ref: '#/components/parameters/LedgerId'
      responses:
        '200':
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
    post:
      tags: [Accounts]
      summary: Create account
      parameters:
        - $ref: '#/components/parameters/LedgerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountCreateRequest'
      responses:
        '201':
          description: Account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
  /ledgers/{ledgerId}/accounts/{accountId}:
    get:
      tags: [Accounts]
      summary: Get account
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
    patch:
      tags: [Accounts]
      summary: Update account
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdateRequest'
      responses:
        '200':
          description: Account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
    delete:
      tags: [Accounts]
      summary: Deactivate account
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - $ref: '#/components/parameters/AccountId'
      responses:
        '204':
          description: Deactivated
  /ledgers/{ledgerId}/categories:
    get:
      tags: [Categories]
      summary: List categories
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - in: query
          name: direction
          schema:
            type: string
            enum: [in, out]
        - in: query
          name: active
          schema:
            type: boolean
      responses:
        '200':
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      tags: [Categories]
      summary: Create category
      parameters:
        - $ref: '#/components/parameters/LedgerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
      responses:
        '201':
          description: Category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
...
  /ledgers/{ledgerId}/transactions:
    post:
      tags: [Journal]
      summary: Create transaction
      parameters:
        - $ref: '#/components/parameters/LedgerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionCreateRequest'
      responses:
        '201':
          description: Transaction
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
    get:
      tags: [Journal]
      summary: List transactions
      parameters:
        - $ref: '#/components/parameters/LedgerId'
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: account_id
          schema:
            type: string
        - in: query
          name: category_id
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: cursor_occurred_at
          schema:
            type: string
            format: date-time
        - in: query
          name: cursor_id
          schema:
            type: string
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionList'
...
  /card-networks:
    get:
      tags: [Cartao]
      summary: List card networks
      responses:
        '200':
          description: Networks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CardNetwork'
    post:
      tags: [Cartao]
      summary: Create card network
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CardNetworkRequest'
      responses:
        '201':
          description: Network
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardNetwork'
...
  /accounts/{accountId}/credit-cards:
    get:
      tags: [Cartao]
      summary: List credit cards for account
      parameters:
        - $ref: '#/components/parameters/AccountId'
      responses:
        '200':
          description: Cards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditCard'
    post:
      tags: [Cartao]
      summary: Create credit card for account
      parameters:
        - $ref: '#/components/parameters/AccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardRequest'
      responses:
        '201':
          description: Credit card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCard'
  /credit-cards/{cardId}:
    get:
      tags: [Cartao]
      summary: Get credit card
      parameters:
        - $ref: '#/components/parameters/CardId'
      responses:
        '200':
          description: Credit card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCard'
    patch:
      tags: [Cartao]
      summary: Update credit card
      parameters:
        - $ref: '#/components/parameters/CardId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardRequest'
      responses:
        '200':
          description: Credit card
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCard'
    delete:
      tags: [Cartao]
      summary: Delete credit card
      parameters:
        - $ref: '#/components/parameters/CardId'
      responses:
        '204':
          description: Deleted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    LedgerId:
      in: path
      name: ledgerId
      required: true
      schema:
        type: string
...
    AccountId:
      in: path
      name: accountId
      required: true
      schema:
        type: string
...
    CardId:
      in: path
      name: cardId
      required: true
      schema:
        type: string
...
  schemas:
...
    Ledger:
      type: object
      properties:
        id:
          type: string
        owner_user_id:
          type: string
        name:
          type: string
        currency_code:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    LedgerWithRole:
      type: object
      properties:
        ledger:
          $ref: '#/components/schemas/Ledger'
        role:
          type: string
...
    LedgerCreateRequest:
      type: object
      properties:
        name:
          type: string
        currency_code:
          type: string
        create_default_accounts:
          type: boolean
        include_investment:
          type: boolean
...
    Account:
      type: object
      properties:
        id:
          type: string
        ledger_id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [current, business, investment, exchange, wallet]
        nature:
          type: string
          enum: [asset, liability]
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    AccountCreateRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [current, business, investment, exchange, wallet]
        nature:
          type: string
          enum: [asset, liability]
        is_active:
          type: boolean
    AccountUpdateRequest:
      type: object
      properties:
        name:
          type: string
        is_active:
          type: boolean
    Category:
      type: object
      properties:
        id:
          type: string
        ledger_id:
          type: string
        parent_id:
          type: string
          nullable: true
        name:
          type: string
        direction:
          type: string
        is_budget_base:
          type: boolean
        is_budget_relevant:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    CategoryCreateRequest:
      type: object
      properties:
        name:
          type: string
        direction:
          type: string
        is_budget_base:
          type: boolean
        is_budget_relevant:
          type: boolean
        parent_id:
          type: string
          nullable: true
...
    Entry:
      type: object
      properties:
        id:
          type: string
        account_id:
          type: string
        category_id:
          type: string
          nullable: true
        kind:
          type: string
        amount_cents:
          type: integer
        memo:
          type: string
          nullable: true
    Transaction:
      type: object
      properties:
        id:
          type: string
        ledger_id:
          type: string
        occurred_at:
          type: string
          format: date-time
        description:
          type: string
        notes:
          type: string
          nullable: true
        created_by_user_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
        entries:
          type: array
          items:
            $ref: '#/components/schemas/Entry'
    TransactionCreateRequest:
      type: object
      properties:
        occurred_at:
          type: string
          format: date-time
        description:
          type: string
        notes:
          type: string
          nullable: true
        entries:
          type: array
          items:
            $ref: '#/components/schemas/Entry'
...
    TransactionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        next_cursor:
          type: object
          properties:
            cursor_occurred_at:
              type: string
              format: date-time
            cursor_id:
              type: string
...
    CardNetwork:
      type: object
      properties:
        code:
          type: string
        display_name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    CardNetworkRequest:
      type: object
      properties:
        code:
          type: string
        display_name:
          type: string
...
    CreditCard:
      type: object
      properties:
        id:
          type: string
        ledger_id:
          type: string
        parent_account_id:
          type: string
        liability_account_id:
          type: string
        label:
          type: string
          nullable: true
        brand:
          type: string
        last4:
          type: string
          nullable: true
        cvv:
          type: string
          nullable: true
        holder_name:
          type: string
          nullable: true
        active:
          type: boolean
        color:
          type: string
          nullable: true
        style:
          type: string
          nullable: true
        closing_day:
          type: integer
        due_day:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    CreditCardRequest:
      type: object
      properties:
        label:
          type: string
        brand:
          type: string
        last4:
          type: string
        cvv:
          type: string
        holder_name:
          type: string
        active:
          type: boolean
        color:
          type: string
        style:
          type: string
        closing_day:
          type: integer
        due_day:
          type: integer
...